medios_pago/pyproject.toml
[project]
name = "coinbase-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.1",
    "fastapi>=0.123.9",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
    "mercadopago>=3.0.0",
    "redis>=5.0.0",
]
